pyproject.toml
[project]
name = "mlops-course"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleh Borysevych",email = "borysevych.oleh87@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "datasets (>=3.6.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "dvc (>=3.59.2,<4.0.0)",
    "dvc-s3 (>=3.2.0,<4.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "fasttext (>=0.9.3,<0.10.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "numpy (==1.*)",
    "pandas (>=2.2.3,<3.0.0)",
    "wandb (>=0.19.11,<0.20.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.black]
line-length = 140

[tool.isort]
py_version = 311
line_length = 140

[tool.flake8]
max-line-length = 140
exclude = [".git", "__pycache__", ".venv"]
extend-ignore = ["B008"]
max-complexity = 8
per-file-ignores = [
    '__init__.py:F401',
]
